+++ layer/decorators/resources_decorator.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Dict, List, Sequence, Union
 
 import wrapt  # type: ignore
 
-from layer.contracts.runs import ResourcePath
+from layer.contracts.remote_runs import ResourcePath
 from layer.decorators.layer_wrapper import LayerFunctionWrapper
