+++ layer/runs/initializer.py
@@ -9,7 +9,7 @@ from layer.clients.layer import LayerClient
 from layer.contracts.fabrics import Fabric
 from layer.contracts.project_full_name import ProjectFullName
 from layer.contracts.projects import Project, ProjectLoader
-from layer.contracts.runs import Run, RunStatus
+from layer.contracts.remote_runs import RemoteRun, RunStatus
 from layer.runs import context
 
 
@@ -84,7 +84,7 @@ class RunInitializer:
 
 
 def finalize_run(
-    layer_client: LayerClient, run_context: context.RunContext, run: Run
+    layer_client: LayerClient, run_context: context.RunContext, run: RemoteRun
 ) -> None:
     """
     This is the atexit handler that sets the status of the run as completed
@@ -115,7 +115,7 @@ def get_or_create_run(
     run_name: Optional[str],
     run_index: Optional[int],
     labels: Optional[Set[str]] = None,
-) -> Run:
+) -> RemoteRun:
     run_id_env = os.getenv(LAYER_RUN_ID_ENV_VARIABLE)
     if run_id_env is not None:
         run = client.run_service_client.get_run_by_id(uuid.UUID(run_id_env))
