+++ test/unit/decorators/test_resources_decorator.py
@@ -2,7 +2,7 @@ import pathlib
 
 import pytest
 
-from layer.contracts.runs import ResourcePath
+from layer.contracts.remote_runs import ResourcePath
 from layer.decorators import resources
